backend/requirements.txt
absl-py==2.3.1
astunparse==1.6.3
blinker==1.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==2.3.3
Flask-Cors==4.0.0
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.1
gunicorn==23.0.0
h5py==3.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.10.0
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.1.0
numpy==1.26.2
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
Pillow==10.1.0
protobuf==4.25.8
Pygments==2.19.2
requests==2.32.4
rich==14.0.0
setuptools==80.9.0
six==1.17.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-cpu==2.16.1
tensorflow-intel==2.16.1
termcolor==3.1.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
